+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased changes can be added here. -->
 - Made content updates for EDPUB 1085, adding 'Sub-Daily'
+- Reduce API calls for request details page
+- Update UMMC metadata mapping template
 
 ## 1.0.19